+++ src/Products/PloneMeeting/content/held_position.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from collective.contact.core import _ as _ccc
 from collective.contact.core.content.held_position import HeldPosition
 from collective.contact.core.content.held_position import IHeldPosition
-from collective.contact.core.schema import ContactChoice
 from collective.contact.core.utils import get_gender_and_number
 from collective.contact.plonegroup.config import PLONEGROUP_ORG
 from collective.excelexport.exportables.dexterityfields import get_exportable_for_fieldname
@@ -12,8 +10,8 @@ from plone.dexterity.schema import DexteritySchemaPolicy
 from plone.supermodel import model
 from Products.CMFPlone.utils import safe_unicode
 from Products.PloneMeeting.config import PMMessageFactory as _
-from Products.PloneMeeting.content.source import PMContactSourceBinder
 from Products.PloneMeeting.utils import plain_render
+from Products.PloneMeeting.utils import uncapitalize
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope.globalrequest import getRequest
 from zope.i18n import translate
@@ -24,16 +22,6 @@ import zope.schema
 
 class IPMHeldPosition(IHeldPosition):
     """ """
-
-    # override position to use a select list restricted to orgs out of PLONEGROUP_ORG
-    position = ContactChoice(
-        title=_ccc("Organization/Position"),
-        description=_("Select an organization, most of time, there is one single organization, "
-                      "search for \"My organization\"."),
-        source=PMContactSourceBinder(),
-        required=True,
-    )
-
     form.order_before(position_type='start_date')
     position_type = zope.schema.Choice(
         title=_("Position type"),
@@ -181,7 +169,12 @@ class PMHeldPosition(HeldPosition):
                    'FP': values[0]}
         return res
 
-    def get_prefix_for_gender_and_number(self, value=None, include_value=False, use_by=False, use_to=False):
+    def get_prefix_for_gender_and_number(self,
+                                         value=None,
+                                         include_value=False,
+                                         include_person_title=False,
+                                         use_by=False,
+                                         use_to=False):
         """Get prefix to use depending on given value."""
         value_starting_vowel = {'MS': u'L\'',
                                 'MP': u'Les ',
@@ -246,6 +239,10 @@ class PMHeldPosition(HeldPosition):
             [self.get_person()], use_by=use_by, use_to=use_to), '') or ''
         if include_value:
             res = u'{0}{1}'.format(res, value)
+        if include_person_title:
+            # we lowerize first letter of res
+            res = uncapitalize(res)
+            res = u'{0} {1}'.format(self.person_title, res)
         return res