+++ src/Products/PloneMeeting/content/organization.py
@@ -2,14 +2,11 @@
 
 from collective.contact.core.content.organization import IOrganization
 from collective.contact.core.content.organization import Organization
-from collective.contact.plonegroup.config import ORGANIZATIONS_REGISTRY
 from collective.contact.plonegroup.config import PLONEGROUP_ORG
 from collective.contact.plonegroup.interfaces import IPloneGroupContact
-from collective.contact.plonegroup.utils import get_organization
 from collective.contact.plonegroup.utils import get_organizations
 from collective.z3cform.datagridfield import DataGridFieldFactory
 from collective.z3cform.datagridfield import DictRow
-from plone import api
 from plone.autoform import directives as form
 from plone.dexterity.schema import DexteritySchemaPolicy
 from plone.supermodel import model
@@ -17,12 +14,10 @@ from Products.PloneMeeting.config import PMMessageFactory as _
 from Products.PloneMeeting.utils import computeCertifiedSignatures
 from Products.PloneMeeting.utils import listifySignatures
 from Products.PloneMeeting.validators import DXCertifiedSignaturesValidator
+from z3c.form import validator
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope import schema
 from zope.interface import Interface
-from zope.interface import Invalid
-from zope.interface import invariant
-from z3c.form import validator
 
 
 class ICertifiedSignaturesRowSchema(Interface):
@@ -184,11 +179,11 @@ class PMOrganization(Organization):
         """Accessor so it can be called in a TAL expression."""
         return self.groups_in_charge
 
-    def get_full_title(self, separator=u' / ', first_index=0):
+    def get_full_title(self, separator=u' / ', first_index=0, force_separator=False):
         """Override to change default first_index from 0 to 1 for IPloneGroupContact,
            so we do not display the 'My organization' level for elements displayed in the
            'My organization' organization."""
-        if self.id != PLONEGROUP_ORG and IPloneGroupContact.providedBy(self):
+        if not force_separator and self.id != PLONEGROUP_ORG and IPloneGroupContact.providedBy(self):
             first_index = 1
         return super(PMOrganization, self).get_full_title(separator, first_index)
 
@@ -246,34 +241,30 @@ class PMOrganization(Organization):
                 res = True
         return res
 
-    def get_certified_signatures(self, computed=False, cfg=None, from_group_in_charge=False, listify=True, **kwargs):
-        '''Overrides field 'certified_signatures' accessor to be able to pass
+    def get_certified_signatures(self, computed=False, cfg=None, group_in_charge=None, listify=True, **kwargs):
+        """Overrides field 'certified_signatures' accessor to be able to pass
            the p_computed parameter that will return computed certified signatures,
            so signatures really available right now.  If nothing is defined on the organization,
            use certified signatures defined on the corresponding p_cfg MeetingConfig.
-           If p_from_group_in_charge is True, we get certifiedSignatures from the first defined
-           self.groupsInCharge.'''
+           If p_from_group_in_charge is an organization, we get certifiedSignatures from it."""
         group_signatures = self.certified_signatures
         if computed:
-            computedSignatures = cfg.getCertifiedSignatures(computed=True)
+            computed_signatures = cfg.getCertifiedSignatures(computed=True)
 
             # get certified signatures from first of the defined groupsInCharge
-            groups_in_charge = self.groups_in_charge
-            if from_group_in_charge and groups_in_charge:
-                group_in_charge = get_organization(groups_in_charge[0])
-                computedSignatures.update(
-                    computeCertifiedSignatures(group_in_charge.get_certified_signatures()))
+            if group_in_charge:
+                computed_signatures.update(computeCertifiedSignatures(group_in_charge.get_certified_signatures()))
 
             # if we have certified signatures defined on this MeetingGroup
             # update MeetingConfig signatures regarding what is defined here
             if group_signatures:
-                computedSignatures.update(computeCertifiedSignatures(group_signatures))
+                computed_signatures.update(computeCertifiedSignatures(group_signatures))
             # listify signatures, for backward compatibility, we need a list of pair
             # of function/name, like ['function1', 'name1', 'function2', 'name2']
             if listify:
-                group_signatures = listifySignatures(computedSignatures)
+                group_signatures = listifySignatures(computed_signatures)
             else:
-                group_signatures = computedSignatures
+                group_signatures = computed_signatures
         return group_signatures
 
     def get_order(self, associated_org_uids=[], cfg=None):