+++ src/Products/PloneMeeting/migrations/migrate_to_4101.py
@@ -3,7 +3,9 @@
 from eea.facetednavigation.interfaces import ICriteria
 from imio.helpers.catalog import reindexIndexes
 from plone import api
+from plone.app.contenttypes.migration.dxmigration import migrate_base_class_to_new_class
 from plone.namedfile.file import NamedBlobFile
+from Products.PloneMeeting.content.organization import PMOrganization
 from Products.PloneMeeting.migrations import logger
 from Products.PloneMeeting.migrations import Migrator
 
@@ -102,6 +104,45 @@ class Migrate_To_4101(Migrator):
                 delattr(org, 'selectable_for_plonegroup')
         logger.info('Done.')
 
+    def _removeTagsParameterInCallToJSCallViewAndReloadInCloneToOtherMCActions(self):
+        '''Parameter was remove from JS callViewAndReload, this is stored
+           in clone to other mc actions on MeetingItem portal_types.'''
+        logger.info("Calling _updatePortalTypes for every MeetingConfigs...")
+        for cfg in self.tool.objectValues('MeetingConfig'):
+            cfg._updatePortalTypes()
+        logger.info('Done.')
+
+    def _moveToMeetingConfigOnMeetingTransitionItemActionToExecute(self):
+        '''Field MeetingConfig.onMeetingTransitionItemTransitionToTrigger was moved
+           to MeetingConfig.onMeetingTransitionItemActionToExecute.'''
+        logger.info("Migrating field onMeetingTransitionItemActionToExecute for every MeetingConfigs...")
+        for cfg in self.tool.objectValues('MeetingConfig'):
+            if hasattr(cfg, 'onMeetingTransitionItemTransitionToTrigger'):
+                old_values = cfg.onMeetingTransitionItemTransitionToTrigger
+                # move 'item_transition' to 'item_action' and add empty 'tal_expression'
+                new_values = []
+                for old_value in old_values:
+                    new_values.append({'meeting_transition': old_value['meeting_transition'],
+                                       'item_action': old_value['item_transition'],
+                                       'tal_expression': ''})
+                cfg.setOnMeetingTransitionItemActionToExecute(new_values)
+                delattr(cfg, 'onMeetingTransitionItemTransitionToTrigger')
+        logger.info('Done.')
+
+    def _migrateContactOrganizationPersonsKlass(self):
+        """klass used by 'organization' portal_type changed, this is only relevant for
+           users using beta versions..."""
+        logger.info('Migrating klass of collective.contact Organization to PMOrganization...')
+        for brain in self.catalog(portal_type='organization'):
+            org = brain.getObject()
+            if not isinstance(org, PMOrganization):
+                logger.info('Migrating organization at {0}'.format('/'.join(org.getPhysicalPath())))
+                migrate_base_class_to_new_class(
+                    org,
+                    old_class_name='collective.contact.core.content.organization.Organization',
+                    new_class_name='Products.PloneMeeting.content.organization.PMOrganization')
+        logger.info('Done.')
+
     def run(self, extra_omitted=[]):
         logger.info('Migrating to PloneMeeting 4101...')
         self._updateFacetedFilters()
@@ -110,8 +151,18 @@ class Migrate_To_4101(Migrator):
         self._allowDashboardPODTemplateInDirectoryPortalType()
         self._addDashboardPODTemplateExportOrganizations()
         self._removeSelectableForPlonegroupFieldOnOrganizations()
+        self._removeTagsParameterInCallToJSCallViewAndReloadInCloneToOtherMCActions()
+        self._moveToMeetingConfigOnMeetingTransitionItemActionToExecute()
+        self._migrateContactOrganizationPersonsKlass()
         self.cleanRegistries()
+        # holidays 2020 were added
+        self.updateHolidays()
+        self.reindexIndexesFor(idxs=['get_full_title'], **{'portal_type': ['organization']})
         reindexIndexes(self.portal, idxs=['getTakenOverBy', 'getConfigId'])
+        # re-run the Meeting workflows update as there was a bug in Migrator.refreshDatabase
+        meeting_wf_ids = self.getWorkflows(meta_types=['Meeting'])
+        self.refreshDatabase(catalogs=False, workflows=True, workflowsToUpdate=meeting_wf_ids)
+        self.tool.invalidateAllCache()
 
 
 def migrate(context):
@@ -123,10 +174,17 @@ def migrate(context):
        4) Add 'DashboardPODTemplate' to the allowed types of a contacts directory;
        5) Add the 'Export CSV' DashboardPODTemplate available on the contacts 'All orgs' dashboard;
        6) Remove the 'selectable_for_plonegroup' attribute on organizations;
-       7) Clean registries;
-       8) Reindex catalog indexes :
+       7) Move field MeetingConfig.onMeetingTransitionItemTransitionToTrigger to
+          MeetingConfig.onMeetingTransitionItemActionToExecute;
+       8) Migrate organizations having klass Organization to klass PMOrganization;
+       9) Clean registries;
+       10) Update holidays to manage 2020;
+       11) Reindex catalog indexes :
               - 'getTakenOverBy' as we use an indexer to handle empty value;
-              - 'getConfigId' as we store a specific empty value as it is not possible to search on an empty index.
+              - 'getConfigId' as we store a specific empty value as it is not possible to search on an empty index;
+              - 'get_full_title' as indexed value is different than displayed one.
+       12) Refresh Meeting workflows so MeetingManager have 'Review portal content' in state 'closed';
+       13) Invalidate all cache.
     '''
     migrator = Migrate_To_4101(context)
     migrator.run()